backend/pyproject.toml
[project]
name = "babel-backend"
version = "0.0.1"
description = "babel — local-first long document translation service"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlmodel>=0.0.22",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.12",
    "httpx>=0.27",
    "tiktoken>=0.8",
    "pymupdf>=1.24",
    "pymupdf4llm>=0.0.17",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12",
    "python-docx>=1.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7",
    "pytest>=8",
    "pytest-asyncio>=0.24",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
